fix connection btween booking and tickets

+++ tickets/db/migrate/20240423184428_create_events.rb
@@ -5,8 +5,8 @@ class CreateEvents < ActiveRecord::Migration[6.1]
       t.date :date
       t.integer :ordinary_ticket_count
       t.integer :vip_ticket_count
-      t.float :ordinary_ticket_price
-      t.float :vip_ticket_price
+      t.integer :ordinary_ticket_price
+      t.integer :vip_ticket_price
 
       t.timestamps
     end
